+++ CHANGELOG.md
@@ -7,6 +7,7 @@ v0.8.1
 *Release date: 2015-01-??*
 
   * Fixed minor bugs
+  * Add visual testing to lettuce tests
 
 v0.8
 ----